Escape Game - A simple game, searching for a key and exit room
I see this game as example to create a game with PowerPoint. In order to get this running using a presentation app on Haiku i create the graphics. I does not found a way to dió this game using GobeProductive, Liblre Office or Calibre, so i start to realise this game with yab.